Network.AWS.ResourceGroupsTagging.GetTagKeys

Description

Returns all tag keys in the specified region for the AWS account.

This operation returns paginated results.

Creating a Request

getTagKeys :: GetTagKeys

Creates a value of GetTagKeys with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtkPaginationToken - A string that indicates that additional data is available. Leave this value empty for your initial request. If the response includes a PaginationToken, use that string for this value to request an additional page of data.

data GetTagKeys

See: getTagKeys smart constructor.

Request Lenses

gtkPaginationToken :: Lens' GetTagKeys (Maybe Text)

A string that indicates that additional data is available. Leave this value empty for your initial request. If the response includes a PaginationToken, use that string for this value to request an additional page of data.

Destructuring the Response

getTagKeysResponse

Arguments

:: Int
gtkrsResponseStatus

-> GetTagKeysResponse

Creates a value of GetTagKeysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtkrsPaginationToken - A string that indicates that the response contains more data than can be returned in a single response. To receive additional data, specify this string for the PaginationToken value in a subsequent request.

gtkrsTagKeys - A list of all tag keys in the AWS account.

gtkrsResponseStatus - -- | The response status code.

data GetTagKeysResponse

See: getTagKeysResponse smart constructor.

Response Lenses

gtkrsPaginationToken :: Lens' GetTagKeysResponse (Maybe Text)

A string that indicates that the response contains more data than can be returned in a single response. To receive additional data, specify this string for the PaginationToken value in a subsequent request.

gtkrsTagKeys :: Lens' GetTagKeysResponse [Text]

A list of all tag keys in the AWS account.

gtkrsResponseStatus :: Lens' GetTagKeysResponse Int

- | The response status code.