Network.AWS.ResourceGroupsTagging.TagResources

Description

Applies one or more tags to the specified resources. Note the following:

Not all resources can have tags. For a list of resources that support tagging, see Supported Resources in the AWS Resource Groups and Tag Editor User Guide .

Each resource can have up to 50 tags. For other limits, see Tag Restrictions in the Amazon EC2 User Guide for Linux Instances .

You can only tag resources that are located in the specified region for the AWS account.

To add tags to a resource, you need the necessary permissions for the service that the resource belongs to as well as permissions for adding tags. For more information, see Obtaining Permissions for Tagging in the AWS Resource Groups and Tag Editor User Guide .

Creating a Request

tagResources

Arguments

:: NonEmpty Text
trResourceARNList

-> TagResources

Creates a value of TagResources with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trResourceARNList - A list of ARNs. An ARN (Amazon Resource Name) uniquely identifies a resource. You can specify a minimum of 1 and a maximum of 20 ARNs (resources) to tag. An ARN can be set to a maximum of 1600 characters. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

trTags - The tags that you want to add to the specified resources. A tag consists of a key and a value that you define.

data TagResources

See: tagResources smart constructor.

Request Lenses

trResourceARNList :: Lens' TagResources (NonEmpty Text)

A list of ARNs. An ARN (Amazon Resource Name) uniquely identifies a resource. You can specify a minimum of 1 and a maximum of 20 ARNs (resources) to tag. An ARN can be set to a maximum of 1600 characters. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

trTags :: Lens' TagResources (HashMap Text Text)

The tags that you want to add to the specified resources. A tag consists of a key and a value that you define.

Destructuring the Response

tagResourcesResponse

Arguments

:: Int
trrsResponseStatus

-> TagResourcesResponse

Creates a value of TagResourcesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trrsFailedResourcesMap - Details of resources that could not be tagged. An error code, status code, and error message are returned for each failed item.

trrsResponseStatus - -- | The response status code.

data TagResourcesResponse

See: tagResourcesResponse smart constructor.

Response Lenses

trrsFailedResourcesMap :: Lens' TagResourcesResponse (HashMap Text FailureInfo)

Details of resources that could not be tagged. An error code, status code, and error message are returned for each failed item.

trrsResponseStatus :: Lens' TagResourcesResponse Int

- | The response status code.