Network.AWS.ResourceGroupsTagging.Types

Description

Service Configuration

resourceGroupsTagging :: Service

API version 2017-01-26 of the Amazon Resource Groups Tagging API SDK configuration.

Errors

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter is missing or a malformed string or invalid or out-of-range value was supplied for the request parameter.

_ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied to limit the frequency of submitted requests.

_PaginationTokenExpiredException :: AsError a => Getting (First ServiceError) a ServiceError

A PaginationToken is valid for a maximum of 15 minutes. Your request was denied because the specified PaginationToken has expired.

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing failed because of an unknown error, exception, or failure. You can retry the request.

ResourceErrorCode

data ResourceErrorCode

Constructors

InternalServiceException

InvalidParameterException

FailureInfo

data FailureInfo

Details of the common errors that all actions return.

See: failureInfo smart constructor.

failureInfo :: FailureInfo

Creates a value of FailureInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fiErrorCode - The code of the common error. Valid values include InternalServiceException , InvalidParameterException , and any valid error code returned by the AWS service that hosts the resource that you want to tag.

fiErrorMessage - The message of the common error.

fiStatusCode - The HTTP status code of the common error.

fiErrorCode :: Lens' FailureInfo (Maybe ResourceErrorCode)

The code of the common error. Valid values include InternalServiceException , InvalidParameterException , and any valid error code returned by the AWS service that hosts the resource that you want to tag.

fiErrorMessage :: Lens' FailureInfo (Maybe Text)

The message of the common error.

fiStatusCode :: Lens' FailureInfo (Maybe Int)

The HTTP status code of the common error.

ResourceTagMapping

data ResourceTagMapping

A list of resource ARNs and the tags (keys and values) that are associated with each.

See: resourceTagMapping smart constructor.

resourceTagMapping :: ResourceTagMapping

Creates a value of ResourceTagMapping with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtmResourceARN - An array of resource ARN(s).

rtmTags - The tags that have been applied to one or more AWS resources.

rtmResourceARN :: Lens' ResourceTagMapping (Maybe Text)

An array of resource ARN(s).

rtmTags :: Lens' ResourceTagMapping [Tag]

The tags that have been applied to one or more AWS resources.

Tag

data Tag

The metadata that you apply to AWS resources to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. For more information, see Tag Basics in the Amazon EC2 User Guide for Linux Instances .

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

tagValue - The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

tagKey :: Lens' Tag Text

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

tagValue :: Lens' Tag Text

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

TagFilter

data TagFilter

A list of tags (keys and values) that are used to specify the associated resources.

See: tagFilter smart constructor.

tagFilter :: TagFilter

Creates a value of TagFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tfValues - The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

tfKey - One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.

tfValues :: Lens' TagFilter [Text]

The optional part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key).

tfKey :: Lens' TagFilter (Maybe Text)

One part of a key-value pair that make up a tag. A key is a general label that acts like a category for more specific tag values.