Network.AWS.ResourceGroupsTagging.UntagResources

Description

Removes the specified tags from the specified resources. When you specify a tag key, the action removes both that key and its associated value. The operation succeeds even if you attempt to remove tags from a resource that were already removed. Note the following:

To remove tags from a resource, you need the necessary permissions for the service that the resource belongs to as well as permissions for removing tags. For more information, see Obtaining Permissions for Tagging in the AWS Resource Groups and Tag Editor User Guide .

You can only tag resources that are located in the specified region for the AWS account.

Creating a Request

untagResources

Arguments

:: NonEmpty Text
urResourceARNList

-> NonEmpty Text
urTagKeys

-> UntagResources

Creates a value of UntagResources with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urResourceARNList - A list of ARNs. An ARN (Amazon Resource Name) uniquely identifies a resource. You can specify a minimum of 1 and a maximum of 20 ARNs (resources) to untag. An ARN can be set to a maximum of 1600 characters. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

urTagKeys - A list of the tag keys that you want to remove from the specified resources.

data UntagResources

See: untagResources smart constructor.

Request Lenses

urResourceARNList :: Lens' UntagResources (NonEmpty Text)

A list of ARNs. An ARN (Amazon Resource Name) uniquely identifies a resource. You can specify a minimum of 1 and a maximum of 20 ARNs (resources) to untag. An ARN can be set to a maximum of 1600 characters. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

urTagKeys :: Lens' UntagResources (NonEmpty Text)

A list of the tag keys that you want to remove from the specified resources.

Destructuring the Response

untagResourcesResponse

Arguments

:: Int
urrsResponseStatus

-> UntagResourcesResponse

Creates a value of UntagResourcesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urrsFailedResourcesMap - Details of resources that could not be untagged. An error code, status code, and error message are returned for each failed item.

urrsResponseStatus - -- | The response status code.

data UntagResourcesResponse

See: untagResourcesResponse smart constructor.

Response Lenses

urrsFailedResourcesMap :: Lens' UntagResourcesResponse (HashMap Text FailureInfo)

Details of resources that could not be untagged. An error code, status code, and error message are returned for each failed item.

urrsResponseStatus :: Lens' UntagResourcesResponse Int

- | The response status code.